@pdyraga 2018-07-26
# 512-bit prime

```
benchmark               old ns/op       new ns/op     delta
BenchmarkPrimeGen-8     2268969681      188363314     -91.70%
BenchmarkPrimeGen-8     11386138267     170493155     -98.50%
BenchmarkPrimeGen-8     2631451953      180446766     -93.14%
BenchmarkPrimeGen-8     1934699051      208869584     -89.20%
BenchmarkPrimeGen-8     10074871723     305999350     -96.96%
BenchmarkPrimeGen-8     5705400096      260922273     -95.43%
BenchmarkPrimeGen-8     5569562094      153497360     -97.24%
BenchmarkPrimeGen-8     1784282054      210098300     -88.23%
BenchmarkPrimeGen-8     1239902622      209132014     -83.13%
BenchmarkPrimeGen-8     2478345181      216459976     -91.27%
BenchmarkPrimeGen-8     3105669082      186708939     -93.99%
BenchmarkPrimeGen-8     2015272750      168540463     -91.64%
BenchmarkPrimeGen-8     2052936393      250378808     -87.80%
BenchmarkPrimeGen-8     2870424928      325760690     -88.65%
BenchmarkPrimeGen-8     6326295939      203200026     -96.79%
BenchmarkPrimeGen-8     6128458193      257232450     -95.80%
BenchmarkPrimeGen-8     1991237139      178546016     -91.03%
BenchmarkPrimeGen-8     4419054637      215965061     -95.11%
BenchmarkPrimeGen-8     1278097111      150693571     -88.21%
BenchmarkPrimeGen-8     1596391885      186586729     -88.31%
BenchmarkPrimeGen-8     2874842860      210377943     -92.68%
BenchmarkPrimeGen-8     2763618377      205416284     -92.57%
BenchmarkPrimeGen-8     1894357212      234921937     -87.60%
BenchmarkPrimeGen-8     2820765637      313129768     -88.90%
BenchmarkPrimeGen-8     5442844899      213699420     -96.07%
BenchmarkPrimeGen-8     2317770026      327961987     -85.85%
BenchmarkPrimeGen-8     13632503186     254280747     -98.13%
BenchmarkPrimeGen-8     1627820128      149842621     -90.79%
BenchmarkPrimeGen-8     1618299835      233881054     -85.55%
BenchmarkPrimeGen-8     1563058438      244266049     -84.37%
BenchmarkPrimeGen-8     6981079384      153383777     -97.80%
BenchmarkPrimeGen-8     8583457685      232863218     -97.29%
BenchmarkPrimeGen-8     2920009879      182719931     -93.74%
BenchmarkPrimeGen-8     2855556611      316134549     -88.93%
BenchmarkPrimeGen-8     4192479325      239415232     -94.29%
BenchmarkPrimeGen-8     2124136761      270599184     -87.26%
BenchmarkPrimeGen-8     3354065514      226163593     -93.26%
BenchmarkPrimeGen-8     4793574227      195853126     -95.91%
BenchmarkPrimeGen-8     7926695778      203830914     -97.43%
BenchmarkPrimeGen-8     3783685714      297339207     -92.14%
BenchmarkPrimeGen-8     5769461617      193390742     -96.65%
BenchmarkPrimeGen-8     3899060159      175458617     -95.50%
BenchmarkPrimeGen-8     8162547820      395694820     -95.15%
BenchmarkPrimeGen-8     1162491142      210804576     -81.87%
BenchmarkPrimeGen-8     1292424634      276726318     -78.59%
BenchmarkPrimeGen-8     2615398761      184457813     -92.95%
BenchmarkPrimeGen-8     3015784825      249611044     -91.72%
BenchmarkPrimeGen-8     2652024361      265722298     -89.98%
BenchmarkPrimeGen-8     2649694963      496443957     -81.26%
BenchmarkPrimeGen-8     1152981067      249966541     -78.32%

benchmark               old allocs     new allocs     delta
BenchmarkPrimeGen-8     362387         377422         +4.15%
BenchmarkPrimeGen-8     1769941        349798         -80.24%
BenchmarkPrimeGen-8     419616         373273         -11.04%
BenchmarkPrimeGen-8     311689         415817         +33.41%
BenchmarkPrimeGen-8     1653901        585598         -64.59%
BenchmarkPrimeGen-8     947327         502208         -46.99%
BenchmarkPrimeGen-8     917727         288982         -68.51%
BenchmarkPrimeGen-8     293204         398834         +36.03%
BenchmarkPrimeGen-8     204671         394669         +92.83%
BenchmarkPrimeGen-8     412135         399917         -2.96%
BenchmarkPrimeGen-8     512568         325195         -36.56%
BenchmarkPrimeGen-8     336663         282991         -15.94%
BenchmarkPrimeGen-8     339021         427862         +26.21%
BenchmarkPrimeGen-8     475326         551846         +16.10%
BenchmarkPrimeGen-8     1045261        355640         -65.98%
BenchmarkPrimeGen-8     1009762        438359         -56.59%
BenchmarkPrimeGen-8     328106         306610         -6.55%
BenchmarkPrimeGen-8     727409         368051         -49.40%
BenchmarkPrimeGen-8     202614         258985         +27.82%
BenchmarkPrimeGen-8     263268         313234         +18.98%
BenchmarkPrimeGen-8     470412         355386         -24.45%
BenchmarkPrimeGen-8     444871         339915         -23.59%
BenchmarkPrimeGen-8     325247         391494         +20.37%
BenchmarkPrimeGen-8     496687         491867         -0.97%
BenchmarkPrimeGen-8     932339         340511         -63.48%
BenchmarkPrimeGen-8     382644         521179         +36.20%
BenchmarkPrimeGen-8     2351353        356008         -84.86%
BenchmarkPrimeGen-8     281696         239016         -15.15%
BenchmarkPrimeGen-8     284146         384130         +35.19%
BenchmarkPrimeGen-8     271630         382155         +40.69%
BenchmarkPrimeGen-8     1205117        199428         -83.45%
BenchmarkPrimeGen-8     1510913        368338         -75.62%
BenchmarkPrimeGen-8     501141         305767         -38.99%
BenchmarkPrimeGen-8     497295         514941         +3.55%
BenchmarkPrimeGen-8     744680         363065         -51.25%
BenchmarkPrimeGen-8     370638         391459         +5.62%
BenchmarkPrimeGen-8     601074         344112         -42.75%
BenchmarkPrimeGen-8     846588         294090         -65.26%
BenchmarkPrimeGen-8     1353487        309463         -77.14%
BenchmarkPrimeGen-8     638008         479816         -24.79%
BenchmarkPrimeGen-8     1012358        318190         -68.57%
BenchmarkPrimeGen-8     688078         281367         -59.11%
BenchmarkPrimeGen-8     1424913        595065         -58.24%
BenchmarkPrimeGen-8     200156         333529         +66.63%
BenchmarkPrimeGen-8     224154         365020         +62.84%
BenchmarkPrimeGen-8     445941         279446         -37.34%
BenchmarkPrimeGen-8     535468         377860         -29.43%
BenchmarkPrimeGen-8     458854         386878         -15.69%
BenchmarkPrimeGen-8     459767         769193         +67.30%
BenchmarkPrimeGen-8     200718         373180         +85.92%

benchmark               old bytes     new bytes     delta
BenchmarkPrimeGen-8     82728176      37112556      -55.14%
BenchmarkPrimeGen-8     409631096     34535683      -91.57%
BenchmarkPrimeGen-8     96872624      36726974      -62.09%
BenchmarkPrimeGen-8     71888744      41056451      -42.89%
BenchmarkPrimeGen-8     378938424     57864424      -84.73%
BenchmarkPrimeGen-8     217741152     49683195      -77.18%
BenchmarkPrimeGen-8     212017904     28595818      -86.51%
BenchmarkPrimeGen-8     67207840      39318799      -41.50%
BenchmarkPrimeGen-8     47274152      38828333      -17.87%
BenchmarkPrimeGen-8     95460120      39496033      -58.63%
BenchmarkPrimeGen-8     118061851     32182768      -72.74%
BenchmarkPrimeGen-8     78798712      27910800      -64.58%
BenchmarkPrimeGen-8     78144912      42148965      -46.06%
BenchmarkPrimeGen-8     108414480     54402728      -49.82%
BenchmarkPrimeGen-8     241962472     35135193      -85.48%
BenchmarkPrimeGen-8     232679504     43332865      -81.38%
BenchmarkPrimeGen-8     74858416      30346100      -59.46%
BenchmarkPrimeGen-8     167936024     36391868      -78.33%
BenchmarkPrimeGen-8     46505936      25568655      -45.02%
BenchmarkPrimeGen-8     60196144      30969552      -48.55%
BenchmarkPrimeGen-8     108738512     35072412      -67.75%
BenchmarkPrimeGen-8     102350984     33572493      -67.20%
BenchmarkPrimeGen-8     74867877      38710753      -48.29%
BenchmarkPrimeGen-8     114624815     48575358      -57.62%
BenchmarkPrimeGen-8     218440752     33697121      -84.57%
BenchmarkPrimeGen-8     86905816      51420408      -40.83%
BenchmarkPrimeGen-8     541723888     35038903      -93.53%
BenchmarkPrimeGen-8     64802296      23532448      -63.69%
BenchmarkPrimeGen-8     64726184      37951417      -41.37%
BenchmarkPrimeGen-8     62328698      37739038      -39.45%
BenchmarkPrimeGen-8     279747264     19709518      -92.95%
BenchmarkPrimeGen-8     347825616     36348116      -89.55%
BenchmarkPrimeGen-8     114803440     30225522      -73.67%
BenchmarkPrimeGen-8     113669288     50717204      -55.38%
BenchmarkPrimeGen-8     172993448     35961875      -79.21%
BenchmarkPrimeGen-8     84498152      38555288      -54.37%
BenchmarkPrimeGen-8     138581872     34002839      -75.46%
BenchmarkPrimeGen-8     193495360     29064355      -84.98%
BenchmarkPrimeGen-8     311015920     30525862      -90.19%
BenchmarkPrimeGen-8     147068656     47471092      -67.72%
BenchmarkPrimeGen-8     234513488     31368265      -86.62%
BenchmarkPrimeGen-8     159871800     27874576      -82.56%
BenchmarkPrimeGen-8     329108176     58733391      -82.15%
BenchmarkPrimeGen-8     45833152      32946302      -28.12%
BenchmarkPrimeGen-8     51570752      36111451      -29.98%
BenchmarkPrimeGen-8     102344696     27662710      -72.97%
BenchmarkPrimeGen-8     123446618     37352648      -69.74%
BenchmarkPrimeGen-8     105297912     38243751      -63.68%
BenchmarkPrimeGen-8     106156028     75890648      -28.51%
BenchmarkPrimeGen-8     45458288      36981435      -18.65%
```