@pdyraga 2018-07-26
# 1024-bit prime

```
benchmark               old ns/op        new ns/op       delta
BenchmarkPrimeGen-8     36861702183      5566762010      -84.90%
BenchmarkPrimeGen-8     45959705486      2908156086      -93.67%
BenchmarkPrimeGen-8     32302799509      1600852073      -95.04%
BenchmarkPrimeGen-8     10559412608      6909158605      -34.57%
BenchmarkPrimeGen-8     26361757373      2518036679      -90.45%
BenchmarkPrimeGen-8     29501857227      1448072197      -95.09%
BenchmarkPrimeGen-8     231272508288     1392535303      -99.40%
BenchmarkPrimeGen-8     7657751078       10036763187     +31.07%
BenchmarkPrimeGen-8     51558997401      8023889630      -84.44%
BenchmarkPrimeGen-8     16745757300      2572674783      -84.64%
BenchmarkPrimeGen-8     28182579779      2706935779      -90.40%
BenchmarkPrimeGen-8     19612730893      1771794157      -90.97%
BenchmarkPrimeGen-8     24840617742      1559935715      -93.72%
BenchmarkPrimeGen-8     2466813963       4869175794      +97.39%
BenchmarkPrimeGen-8     8494698198       5458197106      -35.75%
BenchmarkPrimeGen-8     24406594314      8387214807      -65.64%
BenchmarkPrimeGen-8     111450059549     2289543448      -97.95%
BenchmarkPrimeGen-8     9711480641       1756765315      -81.91%
BenchmarkPrimeGen-8     81936301479      2577900911      -96.85%
BenchmarkPrimeGen-8     2767979308       3723179272      +34.51%
BenchmarkPrimeGen-8     101696370631     3567358343      -96.49%
BenchmarkPrimeGen-8     22492110644      2786609398      -87.61%
BenchmarkPrimeGen-8     84566338764      871273575       -98.97%
BenchmarkPrimeGen-8     17797979625      10846582551     -39.06%
BenchmarkPrimeGen-8     28036722072      1107767565      -96.05%
BenchmarkPrimeGen-8     26167954580      3447014075      -86.83%
BenchmarkPrimeGen-8     16405363943      1506522554      -90.82%
BenchmarkPrimeGen-8     4533869408       3009052149      -33.63%
BenchmarkPrimeGen-8     28292850899      2939190707      -89.61%
BenchmarkPrimeGen-8     58376102081      7330703326      -87.44%
BenchmarkPrimeGen-8     137515732119     2121113847      -98.46%
BenchmarkPrimeGen-8     5343562821       1170653779      -78.09%
BenchmarkPrimeGen-8     4214167930       5996486300      +42.29%
BenchmarkPrimeGen-8     57539380856      4818339856      -91.63%
BenchmarkPrimeGen-8     56463189239      7023259834      -87.56%
BenchmarkPrimeGen-8     31148876342      2628034774      -91.56%
BenchmarkPrimeGen-8     35924957141      10139439977     -71.78%
BenchmarkPrimeGen-8     37037673636      5741866015      -84.50%
BenchmarkPrimeGen-8     123071912422     15779098285     -87.18%
BenchmarkPrimeGen-8     18735582572      1191203409      -93.64%
BenchmarkPrimeGen-8     14950810812      3990970414      -73.31%
BenchmarkPrimeGen-8     45614828915      5893157572      -87.08%
BenchmarkPrimeGen-8     38343598558      2582931037      -93.26%
BenchmarkPrimeGen-8     37089393514      24885103035     -32.91%
BenchmarkPrimeGen-8     21408341771      1005562686      -95.30%
BenchmarkPrimeGen-8     82520362962      12180098905     -85.24%
BenchmarkPrimeGen-8     71017278949      5728216833      -91.93%
BenchmarkPrimeGen-8     34683384212      10680677166     -69.21%
BenchmarkPrimeGen-8     176688401313     1472591237      -99.17%
BenchmarkPrimeGen-8     27576279959      2510611582      -90.90%

benchmark               old allocs     new allocs     delta
BenchmarkPrimeGen-8     1296874        2776508        +114.09%
BenchmarkPrimeGen-8     1629910        1384754        -15.04%
BenchmarkPrimeGen-8     1145560        816307         -28.74%
BenchmarkPrimeGen-8     375806         3260984        +767.73%
BenchmarkPrimeGen-8     936175         1139303        +21.70%
BenchmarkPrimeGen-8     1032492        677510         -34.38%
BenchmarkPrimeGen-8     7860859        664895         -91.54%
BenchmarkPrimeGen-8     278511         4575811        +1542.96%
BenchmarkPrimeGen-8     1772920        3768429        +112.55%
BenchmarkPrimeGen-8     599647         1210124        +101.81%
BenchmarkPrimeGen-8     974381         1260570        +29.37%
BenchmarkPrimeGen-8     684946         796530         +16.29%
BenchmarkPrimeGen-8     895839         692513         -22.70%
BenchmarkPrimeGen-8     84866          2213842        +2508.63%
BenchmarkPrimeGen-8     292856         2442000        +733.86%
BenchmarkPrimeGen-8     865260         3841449        +343.96%
BenchmarkPrimeGen-8     3960309        972024         -75.46%
BenchmarkPrimeGen-8     348814         782942         +124.46%
BenchmarkPrimeGen-8     2872539        1177839        -59.00%
BenchmarkPrimeGen-8     96468          1611606        +1570.61%
BenchmarkPrimeGen-8     3535318        1598072        -54.80%
BenchmarkPrimeGen-8     774395         1221268        +57.71%
BenchmarkPrimeGen-8     2807298        369722         -86.83%
BenchmarkPrimeGen-8     551587         4623520        +738.22%
BenchmarkPrimeGen-8     909261         477328         -47.50%
BenchmarkPrimeGen-8     859315         1338424        +55.75%
BenchmarkPrimeGen-8     549266         617694         +12.46%
BenchmarkPrimeGen-8     155427         1276826        +721.50%
BenchmarkPrimeGen-8     976922         1315785        +34.69%
BenchmarkPrimeGen-8     2063480        3193686        +54.77%
BenchmarkPrimeGen-8     4841641        935303         -80.68%
BenchmarkPrimeGen-8     183380         517693         +182.31%
BenchmarkPrimeGen-8     147059         2375516        +1515.35%
BenchmarkPrimeGen-8     2041437        2040157        -0.06%
BenchmarkPrimeGen-8     1975536        2874974        +45.53%
BenchmarkPrimeGen-8     1074994        1082881        +0.73%
BenchmarkPrimeGen-8     1287402        3817244        +196.51%
BenchmarkPrimeGen-8     1341691        2100462        +56.55%
BenchmarkPrimeGen-8     4409132        6425913        +45.74%
BenchmarkPrimeGen-8     658803         463552         -29.64%
BenchmarkPrimeGen-8     516970         1388957        +168.67%
BenchmarkPrimeGen-8     1628919        2127240        +30.59%
BenchmarkPrimeGen-8     1358365        980612         -27.81%
BenchmarkPrimeGen-8     1300251        9751724        +649.99%
BenchmarkPrimeGen-8     773588         374964         -51.53%
BenchmarkPrimeGen-8     2933710        5006827        +70.67%
BenchmarkPrimeGen-8     2551294        2293989        -10.09%
BenchmarkPrimeGen-8     1139499        4228107        +271.05%
BenchmarkPrimeGen-8     6096582        549261         -90.99%
BenchmarkPrimeGen-8     974002         1009220        +3.62%

benchmark               old bytes      new bytes      delta
BenchmarkPrimeGen-8     410531984      344987328      -15.97%
BenchmarkPrimeGen-8     516871536      172012104      -66.72%
BenchmarkPrimeGen-8     363488520      100463880      -72.36%
BenchmarkPrimeGen-8     118047720      404344976      +242.53%
BenchmarkPrimeGen-8     297614840      141266872      -52.53%
BenchmarkPrimeGen-8     328461424      83951088       -74.44%
BenchmarkPrimeGen-8     2498362152     83168312       -96.67%
BenchmarkPrimeGen-8     88981416       568617328      +539.03%
BenchmarkPrimeGen-8     562466976      470503328      -16.35%
BenchmarkPrimeGen-8     190238184      149378720      -21.48%
BenchmarkPrimeGen-8     309543552      156230640      -49.53%
BenchmarkPrimeGen-8     217364304      98794060       -54.55%
BenchmarkPrimeGen-8     284504936      86219460       -69.69%
BenchmarkPrimeGen-8     26543528       274414312      +933.83%
BenchmarkPrimeGen-8     92278832       304842736      +230.35%
BenchmarkPrimeGen-8     273084064      476264192      +74.40%
BenchmarkPrimeGen-8     1260919112     120177016      -90.47%
BenchmarkPrimeGen-8     110596312      97894424       -11.48%
BenchmarkPrimeGen-8     908471736      146710656      -83.85%
BenchmarkPrimeGen-8     31164144       201423000      +546.33%
BenchmarkPrimeGen-8     1126090384     197493728      -82.46%
BenchmarkPrimeGen-8     245900104      152274684      -38.07%
BenchmarkPrimeGen-8     892851912      45775300       -94.87%
BenchmarkPrimeGen-8     176047536      571741624      +224.77%
BenchmarkPrimeGen-8     289382520      59191272       -79.55%
BenchmarkPrimeGen-8     273224016      168431424      -38.35%
BenchmarkPrimeGen-8     173397688      77421576       -55.35%
BenchmarkPrimeGen-8     48816216       159746704      +227.24%
BenchmarkPrimeGen-8     309511600      162808000      -47.40%
BenchmarkPrimeGen-8     654125952      396621240      -39.37%
BenchmarkPrimeGen-8     1536756272     116150672      -92.44%
BenchmarkPrimeGen-8     57621800       63679392       +10.51%
BenchmarkPrimeGen-8     46709232       295358932      +532.34%
BenchmarkPrimeGen-8     646374992      252268344      -60.97%
BenchmarkPrimeGen-8     624530576      357305168      -42.79%
BenchmarkPrimeGen-8     342230304      135394928      -60.44%
BenchmarkPrimeGen-8     411465472      475151664      +15.48%
BenchmarkPrimeGen-8     427305360      262550304      -38.56%
BenchmarkPrimeGen-8     1403440376     797594248      -43.17%
BenchmarkPrimeGen-8     208987312      56730160       -72.85%
BenchmarkPrimeGen-8     163475552      172827720      +5.72%
BenchmarkPrimeGen-8     516464392      263352608      -49.01%
BenchmarkPrimeGen-8     432335688      122748888      -71.61%
BenchmarkPrimeGen-8     414238240      1212255000     +192.65%
BenchmarkPrimeGen-8     246773984      47038352       -80.94%
BenchmarkPrimeGen-8     927995536      623209176      -32.84%
BenchmarkPrimeGen-8     813041256      284473552      -65.01%
BenchmarkPrimeGen-8     360886104      523694720      +45.11%
BenchmarkPrimeGen-8     1934911896     69402096       -96.41%
BenchmarkPrimeGen-8     309722576      125312600      -59.54%
```